fix bug in bp_read; add Wifi
