+++ cyrw/wifi.go
@@ -1,12 +1,25 @@
 package cyrw
 
 import (
+	"encoding/binary"
+	"errors"
 	"net"
 	"time"
 
 	"github.com/soypat/cyw43439/internal/slog"
 	"github.com/soypat/cyw43439/whd"
 )
+
+func (d *Device) WifiJoin(ssid, password string) (err error) {
+	err = d.set_power_management(PowerSave)
+	if err != nil {
+		return err
+	}
+	if password == "" {
+		return d.join_open(ssid)
+	}
+	return errors.New("not implemented")
+}
 
 func (d *Device) initControl(clm string) error {
 	d.debug("initControl", slog.Int("clm_len", len(clm)))
@@ -81,12 +94,12 @@ func (d *Device) initControl(clm string) error {
 	for i := range evts.events {
 		evts.events[i] = 0xff
 	}
-	evts.Unset(whd.EvRADIO)
-	evts.Unset(whd.EvIF)
-	evts.Unset(whd.EvPROBREQ_MSG)
-	evts.Unset(whd.EvPROBREQ_MSG_RX)
-	evts.Unset(whd.EvPROBRESP_MSG)
-	evts.Unset(whd.EvROAM)
+	evts.Disable(whd.EvRADIO)
+	evts.Disable(whd.EvIF)
+	evts.Disable(whd.EvPROBREQ_MSG)
+	evts.Disable(whd.EvPROBREQ_MSG_RX)
+	evts.Disable(whd.EvPROBRESP_MSG)
+	evts.Disable(whd.EvROAM)
 	buf := make([]byte, evts.Size())
 	evts.Put(buf)
 	d.set_iovar_n("bsscfg:event_msgs", whd.WWD_STA_INTERFACE, buf)
@@ -108,4 +121,72 @@ func (d *Device) initControl(clm string) error {
 
 func (d *Device) MAC() net.HardwareAddr {
 	return net.HardwareAddr(d.mac[:6])
+}
+
+func (d *Device) set_power_management(mode powerManagementMode) error {
+	if !mode.IsValid() {
+		return errors.New("invalid power management mode")
+	}
+	mode_num := mode.mode()
+	if mode_num == 0 {
+		d.set_iovar("pm2_sleep_ret", whd.WWD_STA_INTERFACE, uint32(mode.sleep_ret_ms()))
+		d.set_iovar("bcn_li_bcn", whd.WWD_STA_INTERFACE, uint32(mode.beacon_period()))
+		d.set_iovar("bcn_li_dtim", whd.WWD_STA_INTERFACE, uint32(mode.dtim_period()))
+		d.set_iovar("assoc_listen", whd.WWD_STA_INTERFACE, uint32(mode.assoc()))
+	}
+	return d.set_ioctl(whd.WLC_SET_PM, whd.WWD_STA_INTERFACE, uint32(mode_num))
+}
+
+func (d *Device) join_open(ssid string) error {
+	if len(ssid) > 32 {
+		return errors.New("ssid too long")
+	}
+	const iface = whd.WWD_STA_INTERFACE
+	d.set_iovar("ampdu_ba_wsize", iface, 8)
+	d.set_ioctl(whd.WLC_SET_WSEC, iface, 0)
+	d.set_iovar2("bsscfg:sup_wpa", iface, 0, 0)
+	d.set_ioctl(whd.WLC_SET_INFRA, iface, 1)
+	d.set_ioctl(whd.WLC_SET_AUTH, iface, 0)
+
+	i := ssidInfo{
+		Length: uint32(len(ssid)),
+	}
+	copy(i.SSID[:], ssid)
+
+	return d.wait_for_join(&i)
+}
+
+type ssidInfo struct {
+	Length uint32
+	SSID   [32]byte
+}
+
+func (d *ssidInfo) Put(order binary.ByteOrder, b []byte) {
+	order.PutUint32(b[0:4], d.Length)
+	copy(b[4:36], d.SSID[:min(len(d.SSID), int(d.Length))])
+}
+
+func (d *Device) wait_for_join(ssid *ssidInfo) (err error) {
+	d.eventmask.Enable(whd.EvSET_SSID)
+	d.eventmask.Enable(whd.EvAUTH)
+
+	var buf [36]byte
+	ssid.Put(_busOrder, buf[:])
+
+	err = d.doIoctlSet(whd.WLC_GET_SSID, whd.WWD_STA_INTERFACE, buf[:36])
+	if err != nil {
+		return err
+	}
+
+	// Poll for async events.
+	deadline := time.Now().Add(10 * time.Second)
+	for time.Until(deadline) > 0 {
+		err = d.check_status(d._sendIoctlBuf[:])
+		if err != nil {
+			return err
+		}
+		time.Sleep(time.Second)
+	}
+
+	return nil
 }
